requirements.txt
flask
PyPDF2
python-docx
scikit-learn
pandas
numpy
gunicorn
